refactor: Merge types/docs from http/client.pyi into http/client.py
- Merged type hints and docstrings from `crudclient/http/client.pyi` into `crudclient/http/client.py`.

- Deleted `crudclient/http/client.pyi`.
